% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryx.R
\name{ayx_documentation}
\alias{ayx_documentation}
\title{Output Alteryx tools related information}
\usage{
ayx_documentation(ayx_path, output_xlsx)
}
\arguments{
\item{ayx_path}{the path of Alteryx workflow}

\item{output_xlsx}{the path of output file. Must be a .xlsx file}
}
\description{
Currently only picks up input,output and formulas tools
}
\examples{
\dontrun{
ayx_documentation("myworkflow.yxdb","documentation.xlsx")
}
}
