% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryx.R
\name{ayx_extract_node}
\alias{ayx_extract_node}
\title{Extract information of a simple node AKA:tools}
\usage{
ayx_extract_node(node)
}
\arguments{
\item{node}{}
}
\description{
Extract information of a simple node AKA:tools
}
