% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryx.R
\name{ayx_fl}
\alias{ayx_fl}
\title{Generate formula strings for Alteryx}
\usage{
ayx_fl(x, f = c("+", "-", "*", "/"))
}
\arguments{
\item{x}{a vecotr of column names in Alteryx}

\item{f}{the function will be applied across columns}
}
\value{
a single string which can be copied into Alteryx formula
}
\description{
Genetate a string which can be copied into Alteryx formula tool
}
\examples{
fields <- c("a", "b", "c", "d")
ayx_fl(fields,"+")
}
