% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excel_column_letters}
\alias{excel_column_letters}
\title{All letters of excel column names}
\format{
A named character vector
}
\usage{
excel_column_letters
}
\description{
A vector contains excel column names
}
\keyword{datasets}
