% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_copy2_folder}
\alias{file_copy2_folder}
\title{Copy Files to folders}
\usage{
file_copy2_folder(
  from,
  to,
  affix_type = c("prefix", "suffix"),
  affix,
  from_is_folder = FALSE
)
}
\arguments{
\item{from}{A vector of orgin files's full paths}

\item{to}{folder full paths. can be a vector}

\item{affix_type}{"prefix" or "suffix". Default no affix}

\item{affix}{a character}

\item{from_is_folder}{TRUE or FALSE. Default is FALSE}
}
\description{
Copy Files to folders
}
\examples{
\dontrun{
file_copy2_folder("E:/folder",c("E:/dest1","E:/test2"),affix_type = "suffix",affix = "_copied",from_is_folder = TRUE)
}
}
