% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_cut2_folder}
\alias{file_cut2_folder}
\title{Cut Files to folder}
\usage{
file_cut2_folder(
  from,
  to,
  affix_type = c("prefix", "suffix"),
  affix,
  from_is_folder = FALSE
)
}
\arguments{
\item{from}{A vector of orgin files's full paths}

\item{to}{folder full path. Only length one}

\item{affix_type}{"prefix" or "suffix". Default no affix}

\item{affix}{a character}

\item{from_is_folder}{TRUE or FALSE. Default is FALSE}
}
\description{
Cut Files to folder
}
\examples{
\dontrun{
file_cut2_folder("E:/folder","E:/dest1",affix_type = "suffix",affix = "_copied",from_is_folder = TRUE)
}
}
