% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteryx.R
\name{ayx_documentation}
\alias{ayx_documentation}
\title{Output Alteryx tools related information}
\usage{
ayx_documentation(x, output)
}
\arguments{
\item{x}{the path of Alteryx workflow}

\item{output}{the path of output file. Must be a .xlsx file}
}
\description{
Currently only picks up input,output and formulas tools
}
\examples{
ayx_documentation("myworkflow.yxdb","documentation.xlsx")
}
